Renamed files to reflect the plural of the entity being adressing. Handling of 404 errors on deploy, other styling changes.
